julia 0.7: new open(cmd) semantics
From checking the deprecation code, it seems that

    Base.pipe_reader(proc)

returns the same thing as what it will put in the first element
when assigning

    in_stream, proc = open(cmd)

However, with just that fix, the call to `start_reading` throws
a method error. Things seem to work if I just move that to the
other branch of the if, but this patch should be reviewed by
someone who understands the code better than I do.